pyproject.toml
[project]
name = "lore-agent-task-manager"
version = "0.1.0"
description = "A lightweight CLI task manager for AI coding agents"
readme = "README.md"
license = "MIT"
authors = [{name = "Rodrigo Giacobelli"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0,<9.0",
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/rodrigogiacobelli/lore"

[project.scripts]
lore = "lore.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lore"]
artifacts = ["src/lore/defaults/AGENTS.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "mypy>=1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
